+++ packages/dht/jest.config.js
@@ -1,5 +1,0 @@
-const rootConfig = require('../../jest.config')
-module.exports = {
-    ...rootConfig,
-    setupFilesAfterEnv: rootConfig.setupFilesAfterEnv.concat('./test/utils/customMatchers.ts')
-}