+++ packages/node/test/integration/multiple-publisher-plugins.test.ts
@@ -1,6 +1,5 @@
 import mqtt, { AsyncMqttClient } from 'async-mqtt'
 import WebSocket from 'ws'
-import fetch from 'node-fetch'
 import { StreamPermission } from '@streamr/sdk'
 import { fetchPrivateKeyWithGas, Queue, fastPrivateKey } from '@streamr/test-utils'
 import { Broker } from '../../src/broker'
@@ -130,7 +129,7 @@ describe('multiple publisher plugins', () => {
     })
 
     afterEach(async () => {
-        await broker?.stop()
+        await broker.stop()
     })
 
     it('subscribe by StreamrClient', async () => {