+++ packages/node/test/integration/plugins/operator/announceNodeToStream.test.ts
@@ -1,5 +1,5 @@
 import { _operatorContractUtils } from '@streamr/sdk'
-import { fastPrivateKey } from '@streamr/test-utils'
+import { fastPrivateKey, generateWalletWithGasAndTokens } from '@streamr/test-utils'
 import { collect, toEthereumAddress } from '@streamr/utils'
 import { version as applicationVersion } from '../../../../package.json'
 import { announceNodeToStream } from '../../../../src/plugins/operator/announceNodeToStream'
@@ -12,7 +12,8 @@ describe('announceNodeToStream', () => {
 
     it('publishes to stream', async () => {
         const { operatorContract, nodeWallets } = await _operatorContractUtils.setupOperatorContract({
-            nodeCount: 1
+            nodeCount: 1,
+            generateWalletWithGasAndTokens
         })
         const operatorContractAddress = toEthereumAddress(await operatorContract.getAddress())
         const nodeWallet = nodeWallets[0]