+++ packages/sdk/src/shim/http-https.ts
@@ -1,5 +1,0 @@
-// NB: THIS FILE MUST BE IN ES5
-
-// Browser native fetch implementation does not support the http(s) Agent like node-fetch does
-// eslint-disable-next-line @typescript-eslint/no-extraneous-class
-export class Agent {}