+++ packages/sdk/src/utils/Pipeline.ts
@@ -103,7 +103,7 @@ export class Pipeline<InType, OutType = InType> implements IPipeline<InType, Out
 
     onMessage = Signal.create<[OutType]>()
 
-    onError = ErrorSignal.create<[Error, (InType | OutType)?, number?]>()
+    onError = ErrorSignal.create<[Error, (InType | OutType)?]>()
 
     filter(fn: G.GeneratorFilter<OutType>): Pipeline<InType, OutType> {
         return this.pipe((src) => G.filter(src, fn, this.onError.trigger))