+++ packages/test-utils/jest.config.js
@@ -1,6 +1,0 @@
-const rootConfig = require('../../jest.config')
-
-module.exports = {
-    ...rootConfig,
-    setupFilesAfterEnv: rootConfig.setupFilesAfterEnv.concat('./src/setupCustomMatchers.ts')
-}