+++ packages/trackerless-network/jest.config.js
@@ -1,8 +1,0 @@
-const rootConfig = require('../../jest.config')
-module.exports = {
-    ...rootConfig,
-    setupFilesAfterEnv: rootConfig.setupFilesAfterEnv.concat(
-        '@streamr/test-utils/setupCustomMatchers'
-    ),
-    testTimeout: 15000
-}